self-harvester/requirements.txt
asyncio
requests
python-binance